requirements.txt
Flask>=2.3.0
flask-cors>=4.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
psutil>=5.9.0